password_manager/requirements.txt
asgiref==3.8.1
black==24.8.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
